Move FIE code to fie.js to fix amp4ads-v0.js size regression (#23493)
* Move fie code to fie

* lint

* lint

* fix

* dep-check

* Fix tests.

* Fix circular dependency

* lint

* type checks

* fix circular dependencies

* another one
