+++ test/unit/test-friendly-iframe-embed.js
@@ -14,10 +14,13 @@
  * limitations under the License.
  */
 
+import {BaseElement} from '../../src/base-element';
+import {ElementStub} from '../../src/element-stub';
 import {
   FriendlyIframeEmbed,
   getFriendlyIframeEmbedOptional,
   installFriendlyIframeEmbed,
+  installStandardServicesInEmbed,
   mergeHtmlForTesting,
   setFriendlyIframeEmbedVisible,
   setSrcdocSupportedForTesting,
@@ -25,11 +28,19 @@ import {
 } from '../../src/friendly-iframe-embed';
 import {Services} from '../../src/services';
 import {Signals} from '../../src/utils/signals';
-import {installServiceInEmbedScope} from '../../src/service';
+import {
+  getService,
+  installServiceInEmbedScope,
+  registerServiceBuilder,
+  setParentWindow,
+} from '../../src/service';
+import {installExtensionsService} from '../../src/service/extensions-impl';
 import {isAnimationNone} from '../../testing/test-helper';
 import {layoutRectLtwh} from '../../src/layout-rect';
 import {loadPromise} from '../../src/event-helper';
+import {resetScheduledElementForTesting} from '../../src/service/custom-element-registry';
 import {toggleExperiment} from '../../src/experiments';
+import {updateFieModeForTesting} from '../../src/service/ampdoc-impl';
 
 describes.realWin('friendly-iframe-embed', {amp: true}, env => {
   let window, document;
@@ -37,6 +48,8 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
   let extensionsMock;
   let resourcesMock;
   let ampdocServiceMock;
+  let installExtensionsInChildWindowStub;
+  let installExtensionsInFieStub;
 
   beforeEach(() => {
     window = env.win;
@@ -53,6 +66,14 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
     sandbox.stub(Services, 'ampdocServiceFor').callsFake(() => ampdocService);
 
     iframe = document.createElement('iframe');
+
+    installExtensionsInChildWindowStub = sandbox
+      .stub(FriendlyIframeEmbed.prototype, 'installExtensionsInChildWindow')
+      .returns(Promise.resolve());
+
+    installExtensionsInFieStub = sandbox
+      .stub(FriendlyIframeEmbed.prototype, 'installExtensionsInFie')
+      .returns(Promise.resolve());
   });
 
   afterEach(() => {
@@ -183,20 +204,6 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
       .returns(Promise.resolve())
       .once();
 
-    // Extensions are installed.
-    let installExtDoc;
-    extensionsMock
-      .expects('installExtensionsInFie')
-      .withExactArgs(
-        sinon.match(arg => {
-          installExtDoc = arg;
-          return true;
-        }),
-        ['amp-test'],
-        /* preinstallCallback */ undefined
-      )
-      .once();
-
     let readyResolver = null;
     const readyPromise = new Promise(resolve => {
       readyResolver = resolve;
@@ -214,7 +221,10 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
       .then(embed => {
         expect(childWinForAmpDoc).to.equal(embed.win);
         expect(ampdoc).to.equal(embed.ampdoc);
-        expect(installExtDoc).to.equal(ampdoc);
+        expect(installExtensionsInFieStub).to.be.calledWithMatch(
+          sinon.match.any,
+          sinon.match.same(ampdoc)
+        );
         expect(ampdoc.setReady).to.not.be.called;
         readyResolver();
         return readyPromise;
@@ -267,20 +277,6 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
       .returns(Promise.resolve())
       .once();
 
-    // Extensions are installed.
-    let installExtDoc;
-    extensionsMock
-      .expects('installExtensionsInFie')
-      .withExactArgs(
-        sinon.match(arg => {
-          installExtDoc = arg;
-          return true;
-        }),
-        ['amp-test'],
-        /* preinstallCallback */ undefined
-      )
-      .once();
-
     let readyResolver = null;
     const readyPromise = new Promise(resolve => {
       readyResolver = resolve;
@@ -299,7 +295,10 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
       .then(embed => {
         expect(childWinForAmpDoc).to.equal(embed.win);
         expect(ampdoc).to.equal(embed.ampdoc);
-        expect(installExtDoc).to.equal(ampdoc);
+        expect(installExtensionsInFieStub).to.be.calledWithMatch(
+          sinon.match.any,
+          sinon.match.same(ampdoc)
+        );
         expect(ampdoc.setReady).to.not.be.called;
         readyResolver();
         return readyPromise;
@@ -319,38 +318,21 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
       .returns(Promise.resolve())
       .once();
 
-    // Extensions are installed.
-    let installExtWin;
-    extensionsMock
-      .expects('installExtensionsInChildWindow')
-      .withExactArgs(
-        sinon.match(arg => {
-          installExtWin = arg;
-          return true;
-        }),
-        ['amp-test'],
-        /* preinstallCallback */ undefined
-      )
-      .once();
-
     const embedPromise = installFriendlyIframeEmbed(iframe, document.body, {
       url: 'https://acme.org/url1',
       html: '<amp-test></amp-test>',
       extensionIds: ['amp-test'],
     });
     return embedPromise.then(embed => {
-      expect(installExtWin).to.equal(embed.win);
+      expect(installExtensionsInChildWindowStub).to.be.calledWith(
+        sinon.match.any,
+        sinon.match.same(embed.win)
+      );
     });
   });
 
   it('should pass pre-install callback', () => {
     const preinstallCallback = function() {};
-
-    // Extensions are installed.
-    extensionsMock
-      .expects('installExtensionsInChildWindow')
-      .withExactArgs(sinon.match(() => true), [], preinstallCallback)
-      .once();
 
     const embedPromise = installFriendlyIframeEmbed(
       iframe,
@@ -361,18 +343,20 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
       },
       preinstallCallback
     );
-    return embedPromise;
+    return embedPromise.then(() => {
+      expect(installExtensionsInChildWindowStub).to.be.calledOnce;
+    });
   });
 
   it('should uninstall all resources', () => {
     extensionsMock.expects('preloadExtension').atLeast(1);
-    extensionsMock.expects('installExtensionsInChildWindow').atLeast(1);
     const embedPromise = installFriendlyIframeEmbed(iframe, document.body, {
       url: 'https://acme.org/url1',
       html: '<amp-test></amp-test>',
       extensionIds: ['amp-test'],
     });
     return embedPromise.then(embed => {
+      expect(installExtensionsInChildWindowStub).to.be.calledOnce;
       resourcesMock
         .expects('removeForChildWindow')
         .withExactArgs(embed.win)
@@ -427,12 +411,9 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
       .returns(Promise.resolve())
       .once();
 
-    // Extensions are installed.
-    extensionsMock.expects('installExtensionsInFie').once();
-
     sandbox
       .stub(FriendlyIframeEmbed.prototype, 'whenReady')
-      .callsFake(() => Promise.resolve());
+      .returns(Promise.resolve());
 
     const embedPromise = installFriendlyIframeEmbed(iframe, document.body, {
       url: 'https://acme.org/url1',
@@ -441,6 +422,7 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
     });
     return embedPromise
       .then(embed => {
+        expect(installExtensionsInFieStub).to.be.calledOnce;
         embed.destroy();
       })
       .then(() => {
@@ -449,13 +431,13 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
   });
 
   it('should start invisible by default and update on request', () => {
-    extensionsMock.expects('installExtensionsInChildWindow').once();
     const embedPromise = installFriendlyIframeEmbed(iframe, document.body, {
       url: 'https://acme.org/url1',
       html: '',
       extensionIds: [],
     });
     return embedPromise.then(embed => {
+      expect(installExtensionsInChildWindowStub).to.be.calledOnce;
       expect(embed.isVisible()).to.be.false;
       const spy = sandbox.spy();
       embed.onVisibilityChanged(spy);
@@ -820,7 +802,6 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
         services: {
           'extensions': {
             obj: {
-              installExtensionsInChildWindow: () => {},
               preloadExtension: () => {},
             },
           },
@@ -1116,6 +1097,536 @@ describes.realWin('friendly-iframe-embed', {amp: true}, env => {
       expect(iframe.style.bottom).to.be.empty;
       expect(iframe.style.width).to.be.empty;
       expect(iframe.style.height).to.be.empty;
+    });
+  });
+});
+
+class AmpTest extends BaseElement {}
+class AmpTestSub extends BaseElement {}
+
+// TODO(#22733): remove once ampdoc-fie is launched.
+describes.realWin('installExtensionsInChildWindow', {amp: true}, env => {
+  let parentWin;
+  let extensions;
+  let extensionsMock;
+  let iframe;
+  let iframeWin;
+  let fie;
+
+  beforeEach(() => {
+    parentWin = env.win;
+    resetScheduledElementForTesting(parentWin, 'amp-test');
+    installExtensionsService(parentWin);
+    extensions = Services.extensionsFor(parentWin);
+    extensionsMock = sandbox.mock(extensions);
+
+    [
+      'urlForDoc',
+      'actionServiceForDoc',
+      'standardActionsForDoc',
+      'navigationForDoc',
+      'timerFor',
+    ].forEach(s => {
+      class FakeService {
+        static installInEmbedWindow() {}
+      }
+      sandbox.stub(FakeService, 'installInEmbedWindow');
+      sandbox.stub(Services, s).returns(new FakeService());
+    });
+
+    iframe = parentWin.document.createElement('iframe');
+    const promise = loadPromise(iframe);
+    const html = '<div id="one"></div>';
+    if ('srcdoc' in iframe) {
+      iframe.srcdoc = html;
+    } else {
+      iframe.src = 'about:blank';
+      const childDoc = iframe.contentWindow.document;
+      childDoc.open();
+      childDoc.write(html);
+      childDoc.close();
+    }
+    parentWin.document.body.appendChild(iframe);
+    fie = new FriendlyIframeEmbed(
+      iframe,
+      {
+        url: 'https://acme.org/url1',
+        html: '<body></body>',
+      },
+      Promise.resolve(),
+      env.ampdoc
+    );
+    return promise.then(() => {
+      iframeWin = iframe.contentWindow;
+      setParentWindow(iframeWin, parentWin);
+    });
+  });
+
+  afterEach(() => {
+    if (iframe.parentElement) {
+      iframe.parentElement.removeChild(iframe);
+    }
+    extensionsMock.verify();
+  });
+
+  it('should set window hierarchy', () => {
+    fie.installExtensionsInChildWindow(extensions, iframeWin, []);
+    expect(iframeWin.__AMP_PARENT).to.equal(parentWin);
+    expect(iframeWin.__AMP_TOP).to.equal(parentWin);
+  });
+
+  it('should install runtime styles', () => {
+    fie.installExtensionsInChildWindow(extensions, iframeWin, []);
+    expect(iframeWin.document.querySelector('style[amp-runtime]')).to.exist;
+  });
+
+  it('should install built-ins', () => {
+    fie.installExtensionsInChildWindow(extensions, iframeWin, []);
+    expect(iframeWin.ampExtendedElements).to.exist;
+    expect(iframeWin.ampExtendedElements['amp-img']).to.exist;
+    expect(iframeWin.ampExtendedElements['amp-img']).to.not.equal(ElementStub);
+    expect(iframeWin.ampExtendedElements['amp-pixel']).to.exist;
+    expect(iframeWin.ampExtendedElements['amp-pixel']).to.not.equal(
+      ElementStub
+    );
+    // Legacy elements are installed as well.
+    expect(iframeWin.ampExtendedElements['amp-ad']).to.equal(ElementStub);
+    expect(iframeWin.ampExtendedElements['amp-embed']).to.equal(ElementStub);
+    expect(iframeWin.ampExtendedElements['amp-video']).to.equal(ElementStub);
+  });
+
+  it('should adopt standard services', () => {
+    fie.installExtensionsInChildWindow(extensions, iframeWin, []);
+
+    const any = {}; // Input doesn't matter since services are stubbed.
+    const url = Services.urlForDoc(any);
+    const actions = Services.actionServiceForDoc(any);
+    const standardActions = Services.standardActionsForDoc(any);
+    const navigation = Services.navigationForDoc(any);
+
+    expect(url.constructor.installInEmbedWindow).to.be.called;
+    expect(actions.constructor.installInEmbedWindow).to.be.called;
+    expect(standardActions.constructor.installInEmbedWindow).to.be.called;
+    expect(navigation.constructor.installInEmbedWindow).to.be.called;
+
+    expect(getService(iframeWin, 'timer')).to.exist;
+  });
+
+  it('should install extensions in child window', () => {
+    const extHolder = extensions.getExtensionHolder_('amp-test');
+    extHolder.scriptPresent = true;
+    const promise = fie.installExtensionsInChildWindow(extensions, iframeWin, [
+      'amp-test',
+    ]);
+    // Must be stubbed already.
+    expect(iframeWin.ampExtendedElements['amp-test']).to.equal(ElementStub);
+    expect(
+      iframeWin.document.createElement('amp-test').implementation_
+    ).to.be.instanceOf(ElementStub);
+    expect(iframeWin.ampExtendedElements['amp-test-sub']).to.be.undefined;
+    // Resolve the promise.
+    extensions.registerExtension(
+      'amp-test',
+      AMP => {
+        // Main extension with CSS.
+        AMP.registerElement('amp-test', AmpTest, 'a{}');
+        // Secondary extension w/o CSS.
+        AMP.registerElement('amp-test-sub', AmpTestSub);
+      },
+      parentWin.AMP
+    );
+    return promise.then(() => {
+      // Main extension.
+      expect(parentWin.ampExtendedElements['amp-test']).to.be.undefined;
+      expect(iframeWin.ampExtendedElements['amp-test']).to.equal(AmpTest);
+      expect(iframeWin.document.querySelector('style[amp-extension=amp-test]'))
+        .to.exist;
+      // Must be upgraded already.
+      expect(
+        iframeWin.document.createElement('amp-test').implementation_
+      ).to.be.instanceOf(AmpTest);
+
+      // Secondary extension.
+      expect(parentWin.ampExtendedElements['amp-test-sub']).to.be.undefined;
+      expect(iframeWin.ampExtendedElements['amp-test-sub']).to.equal(
+        AmpTestSub
+      );
+      expect(
+        iframeWin.document.querySelector('style[amp-extension=amp-test-sub]')
+      ).to.not.exist;
+      // Must be upgraded already.
+      expect(
+        iframeWin.document.createElement('amp-test-sub').implementation_
+      ).to.be.instanceOf(AmpTestSub);
+    });
+  });
+
+  it('should adopt extension services', () => {
+    class FooService {
+      static installInEmbedWindow() {}
+    }
+    sandbox.stub(FooService, 'installInEmbedWindow');
+    registerServiceBuilder(
+      parentWin,
+      'fake-service-foo',
+      FooService,
+      /* opt_instantiate */ true
+    );
+
+    class BarService {
+      static installInEmbedWindow() {}
+    }
+    sandbox.stub(BarService, 'installInEmbedWindow');
+    registerServiceBuilder(
+      parentWin,
+      'fake-service-bar',
+      BarService,
+      /* opt_instantiate */ true
+    );
+
+    const extHolder = extensions.getExtensionHolder_('amp-test');
+    extHolder.scriptPresent = true;
+    const install = fie.installExtensionsInChildWindow(extensions, iframeWin, [
+      'amp-test',
+    ]);
+
+    // Resolve the promise `install`.
+    extensions.registerExtension(
+      'amp-test',
+      AMP => {
+        AMP.registerServiceForDoc('fake-service-foo', FooService);
+      },
+      parentWin.AMP
+    );
+
+    return install.then(() => {
+      expect(FooService.installInEmbedWindow).calledOnce;
+      expect(BarService.installInEmbedWindow).to.not.be.called;
+    });
+  });
+
+  // TODO(#16916): Make this test work with synchronous throws.
+  it.skip('should call pre-install callback before other installs', () => {
+    let preinstallCount = 0;
+    const extHolder = extensions.getExtensionHolder_('amp-test');
+    extHolder.scriptPresent = true;
+    const promise = fie.installExtensionsInChildWindow(
+      extensions,
+      iframeWin,
+      ['amp-test'],
+      function() {
+        // Built-ins not installed yet.
+        expect(
+          iframeWin.ampExtendedElements &&
+            iframeWin.ampExtendedElements['amp-img']
+        ).to.not.exist;
+        // Extension is not loaded yet.
+        expect(
+          iframeWin.ampExtendedElements &&
+            iframeWin.ampExtendedElements['amp-test']
+        ).to.not.exist;
+        preinstallCount++;
+      }
+    );
+    expect(preinstallCount).to.equal(1);
+    expect(iframeWin.ampExtendedElements).to.exist;
+    expect(iframeWin.ampExtendedElements['amp-img']).to.exist;
+    expect(iframeWin.ampExtendedElements['amp-img']).to.not.equal(ElementStub);
+    expect(iframeWin.ampExtendedElements['amp-test']).to.equal(ElementStub);
+
+    // Resolve the promise.
+    extensions.registerExtension(
+      'amp-test',
+      AMP => {
+        AMP.registerElement('amp-test', AmpTest);
+      },
+      parentWin.AMP
+    );
+    return promise.then(() => {
+      // Extension elements are stubbed immediately, but registered only
+      // after extension is loaded.
+      expect(iframeWin.ampExtendedElements['amp-test']).to.equal(AmpTest);
+    });
+  });
+
+  describe('installStandardServicesInEmbed', () => {
+    it('verify order of adopted services for embed', () => {
+      installStandardServicesInEmbed(iframeWin);
+
+      const any = {}; // Input doesn't matter since services are stubbed.
+      const url = Services.urlForDoc(any);
+      const actions = Services.actionServiceForDoc(any);
+      const standardActions = Services.standardActionsForDoc(any);
+      const navigation = Services.navigationForDoc(any);
+
+      // Expected order: url, action, standard-actions, navigation, timer.
+      const one = url.constructor.installInEmbedWindow;
+      const two = actions.constructor.installInEmbedWindow;
+      const three = standardActions.constructor.installInEmbedWindow;
+      const four = navigation.constructor.installInEmbedWindow;
+
+      expect(one).to.be.calledBefore(two);
+      expect(two).to.be.calledBefore(three);
+      expect(three).to.be.calledBefore(four);
+      expect(four).to.be.called;
+    });
+  });
+});
+
+describes.realWin('installExtensionsInFie', {amp: true}, env => {
+  let parentWin;
+  let extensions;
+  let extensionsMock;
+  let iframe;
+  let iframeWin, iframeDocEl;
+  let ampdoc;
+  let fie;
+
+  beforeEach(() => {
+    parentWin = env.win;
+    toggleExperiment(parentWin, 'ampdoc-fie', true);
+    resetScheduledElementForTesting(parentWin, 'amp-test');
+    installExtensionsService(parentWin);
+    extensions = Services.extensionsFor(parentWin);
+    extensionsMock = sandbox.mock(extensions);
+    const ampdocService = Services.ampdocServiceFor(parentWin);
+    updateFieModeForTesting(ampdocService, true);
+
+    iframe = parentWin.document.createElement('iframe');
+    const promise = loadPromise(iframe);
+    const html = '<div id="one"></div>';
+    if ('srcdoc' in iframe) {
+      iframe.srcdoc = html;
+    } else {
+      iframe.src = 'about:blank';
+      const childDoc = iframe.contentWindow.document;
+      childDoc.open();
+      childDoc.write(html);
+      childDoc.close();
+    }
+    parentWin.document.body.appendChild(iframe);
+    fie = new FriendlyIframeEmbed(
+      iframe,
+      {
+        url: 'https://acme.org/url1',
+        html: '<body></body>',
+      },
+      Promise.resolve(),
+      env.ampdoc
+    );
+    return promise.then(() => {
+      iframeWin = iframe.contentWindow;
+      iframeDocEl = iframeWin.document.documentElement;
+      setParentWindow(iframeWin, parentWin);
+      ampdoc = ampdocService.installFieDoc(
+        'https://example.com/embed',
+        iframeWin
+      );
+    });
+  });
+
+  afterEach(() => {
+    toggleExperiment(parentWin, 'ampdoc-fie', false);
+    if (iframe.parentElement) {
+      iframe.parentElement.removeChild(iframe);
+    }
+    extensionsMock.verify();
+  });
+
+  it('should set window hierarchy', () => {
+    fie.installExtensionsInFie(extensions, ampdoc, []);
+    expect(iframeWin.__AMP_PARENT).to.equal(parentWin);
+    expect(iframeWin.__AMP_TOP).to.equal(parentWin);
+  });
+
+  it('should install runtime styles', () => {
+    fie.installExtensionsInFie(extensions, ampdoc, []);
+    expect(iframeWin.document.querySelector('style[amp-runtime]')).to.exist;
+  });
+
+  it('should install built-ins', () => {
+    fie.installExtensionsInFie(extensions, ampdoc, []);
+    expect(iframeWin.ampExtendedElements).to.exist;
+    expect(iframeWin.ampExtendedElements['amp-img']).to.exist;
+    expect(iframeWin.ampExtendedElements['amp-img']).to.not.equal(ElementStub);
+    expect(iframeWin.ampExtendedElements['amp-pixel']).to.exist;
+    expect(iframeWin.ampExtendedElements['amp-pixel']).to.not.equal(
+      ElementStub
+    );
+    // Legacy elements are installed as well.
+    expect(iframeWin.ampExtendedElements['amp-ad']).to.equal(ElementStub);
+    expect(iframeWin.ampExtendedElements['amp-embed']).to.equal(ElementStub);
+    expect(iframeWin.ampExtendedElements['amp-video']).to.equal(ElementStub);
+  });
+
+  it('should create new standard services', () => {
+    fie.installExtensionsInFie(extensions, ampdoc, []);
+
+    const url = Services.urlForDoc(iframeDocEl);
+    const actions = Services.actionServiceForDoc(iframeDocEl);
+    const standardActions = Services.standardActionsForDoc(iframeDocEl);
+    const navigation = Services.navigationForDoc(iframeDocEl);
+
+    expect(url).to.exist;
+    expect(actions).to.exist;
+    expect(standardActions).to.exist;
+    expect(navigation).to.exist;
+    expect(Services.timerFor(iframeWin)).to.exist;
+
+    const parentUrl = Services.urlForDoc(parentWin.document.head);
+    expect(parentUrl).to.exist;
+    expect(url).to.not.equal(parentUrl);
+  });
+
+  it('should adopt parent standard services', () => {
+    fie.installExtensionsInFie(extensions, ampdoc, []);
+
+    const viewer = Services.urlForDoc(iframeDocEl);
+    const parentViewer = Services.urlForDoc(parentWin.document.head);
+    expect(viewer).to.exist;
+    expect(parentViewer).to.exist;
+    expect(viewer).to.not.equal(parentViewer);
+  });
+
+  it('should install extensions in child window', () => {
+    const extHolder = extensions.getExtensionHolder_('amp-test');
+    extHolder.scriptPresent = true;
+    const promise = fie.installExtensionsInFie(extensions, ampdoc, [
+      'amp-test',
+    ]);
+    // Must be stubbed already.
+    expect(iframeWin.ampExtendedElements['amp-test']).to.equal(ElementStub);
+    expect(
+      iframeWin.document.createElement('amp-test').implementation_
+    ).to.be.instanceOf(ElementStub);
+    expect(iframeWin.ampExtendedElements['amp-test-sub']).to.be.undefined;
+    // Resolve the promise.
+    extensions.registerExtension(
+      'amp-test',
+      AMP => {
+        // Main extension with CSS.
+        AMP.registerElement('amp-test', AmpTest, 'a{}');
+        // Secondary extension w/o CSS.
+        AMP.registerElement('amp-test-sub', AmpTestSub);
+      },
+      parentWin.AMP
+    );
+    return promise.then(() => {
+      // Main extension.
+      expect(parentWin.ampExtendedElements['amp-test']).to.be.undefined;
+      expect(iframeWin.ampExtendedElements['amp-test']).to.equal(AmpTest);
+      expect(iframeWin.document.querySelector('style[amp-extension=amp-test]'))
+        .to.exist;
+      // Must be upgraded already.
+      expect(
+        iframeWin.document.createElement('amp-test').implementation_
+      ).to.be.instanceOf(AmpTest);
+
+      // Secondary extension.
+      expect(parentWin.ampExtendedElements['amp-test-sub']).to.be.undefined;
+      expect(iframeWin.ampExtendedElements['amp-test-sub']).to.equal(
+        AmpTestSub
+      );
+      expect(
+        iframeWin.document.querySelector('style[amp-extension=amp-test-sub]')
+      ).to.not.exist;
+      // Must be upgraded already.
+      expect(
+        iframeWin.document.createElement('amp-test-sub').implementation_
+      ).to.be.instanceOf(AmpTestSub);
+    });
+  });
+
+  it('should adopt extension services', () => {
+    const fooConstructorSpy = sandbox.spy();
+    class FooService {
+      constructor(arg) {
+        fooConstructorSpy(arg);
+      }
+    }
+    registerServiceBuilder(
+      parentWin,
+      'fake-service-foo',
+      FooService,
+      /* opt_instantiate */ false
+    );
+
+    const barConstructorSpy = sandbox.spy();
+    class BarService {
+      constructor(arg) {
+        barConstructorSpy(arg);
+      }
+    }
+    registerServiceBuilder(
+      parentWin,
+      'fake-service-bar',
+      BarService,
+      /* opt_instantiate */ false
+    );
+
+    const extHolder = extensions.getExtensionHolder_('amp-test');
+    extHolder.scriptPresent = true;
+    const install = fie.installExtensionsInFie(extensions, ampdoc, [
+      'amp-test',
+    ]);
+
+    // Resolve the promise `install`.
+    extensions.registerExtension(
+      'amp-test',
+      AMP => {
+        AMP.registerServiceForDoc('fake-service-foo', FooService);
+      },
+      parentWin.AMP
+    );
+
+    return install.then(() => {
+      expect(fooConstructorSpy).calledOnce.calledWith(ampdoc);
+      expect(barConstructorSpy).to.not.be.called;
+    });
+  });
+
+  it('should call pre-install callback before other installs', () => {
+    let preinstallCount = 0;
+    const extHolder = extensions.getExtensionHolder_('amp-test');
+    extHolder.scriptPresent = true;
+    const promise = fie.installExtensionsInFie(
+      extensions,
+      ampdoc,
+      ['amp-test'],
+      function(winArg, ampdocArg) {
+        expect(winArg).to.equal(iframeWin);
+        expect(ampdocArg).to.equal(ampdoc);
+        // Built-ins not installed yet.
+        expect(
+          iframeWin.ampExtendedElements &&
+            iframeWin.ampExtendedElements['amp-img']
+        ).to.not.exist;
+        // Extension is not loaded yet.
+        expect(
+          iframeWin.ampExtendedElements &&
+            iframeWin.ampExtendedElements['amp-test']
+        ).to.not.exist;
+        preinstallCount++;
+      }
+    );
+    expect(preinstallCount).to.equal(1);
+    expect(iframeWin.ampExtendedElements).to.exist;
+    expect(iframeWin.ampExtendedElements['amp-img']).to.exist;
+    expect(iframeWin.ampExtendedElements['amp-img']).to.not.equal(ElementStub);
+    expect(iframeWin.ampExtendedElements['amp-test']).to.equal(ElementStub);
+
+    // Resolve the promise.
+    extensions.registerExtension(
+      'amp-test',
+      AMP => {
+        AMP.registerElement('amp-test', AmpTest);
+      },
+      parentWin.AMP
+    );
+    return promise.then(() => {
+      // Extension elements are stubbed immediately, but registered only
+      // after extension is loaded.
+      expect(iframeWin.ampExtendedElements['amp-test']).to.equal(AmpTest);
     });
   });
 });
